consistent-sampling: rebase on upstream ComposableSampler API
Rewrites the consistent-sampling module on top of the composable-samplers API in opentelemetry-sdk-extension-incubator. The legacy OTEP 4673 probability samplers are removed, the consistent56 package is promoted into io.opentelemetry.contrib.sampler.consistent, and classes that duplicated upstream functionality are dropped. The remaining ConsistentRateLimitingSampler, ConsistentVariableThresholdSampler, and ConsistentAnyOf now implement upstream ComposableSampler. ConsistentReservoirSamplingSpanProcessor and the parentbased_consistent_probability autoconfigure provider are retained; the provider now delegates to CompositeSampler.wrap(ComposableSampler.parentThreshold(ComposableSampler.probability(p))).
