--- consistent-sampling/src/main/java/io/opentelemetry/contrib/sampler/consistent56/ConsistentThresholdSampler.java
@@ -1,60 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.contrib.sampler.consistent56;
-
-import static io.opentelemetry.contrib.sampler.consistent56.ConsistentSamplingUtil.appendLast56BitHexEncodedWithoutTrailingZeros;
-import static io.opentelemetry.contrib.sampler.consistent56.ConsistentSamplingUtil.calculateSamplingProbability;
-import static io.opentelemetry.contrib.sampler.consistent56.ConsistentSamplingUtil.checkThreshold;
-import static io.opentelemetry.contrib.sampler.consistent56.ConsistentSamplingUtil.getInvalidThreshold;
-import static io.opentelemetry.contrib.sampler.consistent56.ConsistentSamplingUtil.getMaxThreshold;
-
-import io.opentelemetry.api.common.Attributes;
-import io.opentelemetry.api.trace.SpanKind;
-import io.opentelemetry.context.Context;
-import io.opentelemetry.sdk.trace.data.LinkData;
-import java.util.List;
-
-public abstract class ConsistentThresholdSampler extends ConsistentSampler {
-
-  protected abstract long getThreshold();
-
-  protected static long getThreshold(long threshold) {
-    checkThreshold(threshold);
-    return threshold;
-  }
-
-  protected static String getThresholdDescription(long threshold) {
-    String thresholdString;
-    if (threshold == getMaxThreshold()) {
-      thresholdString = "max";
-    } else {
-      thresholdString =
-          appendLast56BitHexEncodedWithoutTrailingZeros(new StringBuilder(), threshold).toString();
-    }
-
-    return "ConsistentFixedThresholdSampler{threshold="
-        + thresholdString
-        + ", sampling probability="
-        + calculateSamplingProbability(threshold)
-        + "}";
-  }
-
-  @Override
-  public SamplingIntent getSamplingIntent(
-      Context parentContext,
-      String name,
-      SpanKind spanKind,
-      Attributes attributes,
-      List<LinkData> parentLinks) {
-
-    return () -> {
-      if (getThreshold() == getMaxThreshold()) {
-        return getInvalidThreshold();
-      }
-      return getThreshold();
-    };
-  }
-}
