+++ consistent-sampling/src/test/java/io/opentelemetry/contrib/sampler/consistent/TestUtil.java
@@ -5,73 +5,26 @@
 
 package io.opentelemetry.contrib.sampler.consistent;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import java.util.SplittableRandom;
 
-import java.util.HashMap;
-import java.util.Map;
-import org.hipparchus.stat.inference.GTest;
+final class TestUtil {
 
-public final class TestUtil {
+  private static final char[] HEX_DIGITS = {
+    '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f'
+  };
 
   private TestUtil() {}
 
-  public static void verifyObservedPvaluesUsingGtest(
-      long originalNumberOfSpans, Map<Integer, Long> observedPvalues, double samplingProbability) {
-
-    Object notSampled =
-        new Object() {
-          @Override
-          public String toString() {
-            return "NOT SAMPLED";
-          }
-        };
-
-    Map<Object, Double> expectedProbabilities = new HashMap<>();
-    if (samplingProbability >= 1.) {
-      expectedProbabilities.put(0, 1.);
-    } else if (samplingProbability <= 0.) {
-      expectedProbabilities.put(notSampled, 1.);
-    } else {
-      int exponent = 0;
-      while (true) {
-        if (Math.pow(0.5, exponent + 1) < samplingProbability
-            && Math.pow(0.5, exponent) >= samplingProbability) {
-          break;
-        }
-        exponent += 1;
-      }
-      if (samplingProbability == Math.pow(0.5, exponent)) {
-        expectedProbabilities.put(notSampled, 1 - samplingProbability);
-        expectedProbabilities.put(exponent, samplingProbability);
-      } else {
-        expectedProbabilities.put(notSampled, 1 - samplingProbability);
-        expectedProbabilities.put(exponent, 2 * samplingProbability - Math.pow(0.5, exponent));
-        expectedProbabilities.put(exponent + 1, Math.pow(0.5, exponent) - samplingProbability);
-      }
+  static String generateRandomTraceId(SplittableRandom random) {
+    StringBuilder sb = new StringBuilder(32);
+    long hi = random.nextLong();
+    long lo = random.nextLong();
+    for (int i = 0; i < 64; i += 4) {
+      sb.append(HEX_DIGITS[(int) (hi >>> i) & 0xF]);
     }
-
-    Map<Object, Long> extendedObservedAdjustedCounts = new HashMap<>(observedPvalues);
-    long numberOfSpansNotSampled =
-        originalNumberOfSpans - observedPvalues.values().stream().mapToLong(i -> i).sum();
-    if (numberOfSpansNotSampled > 0) {
-      extendedObservedAdjustedCounts.put(notSampled, numberOfSpansNotSampled);
+    for (int i = 0; i < 64; i += 4) {
+      sb.append(HEX_DIGITS[(int) (lo >>> i) & 0xF]);
     }
-
-    double[] expectedValues = new double[expectedProbabilities.size()];
-    long[] observedValues = new long[expectedProbabilities.size()];
-
-    int counter = 0;
-    for (Object key : expectedProbabilities.keySet()) {
-      observedValues[counter] = extendedObservedAdjustedCounts.getOrDefault(key, 0L);
-      double p = expectedProbabilities.get(key);
-      expectedValues[counter] = p * originalNumberOfSpans;
-      counter += 1;
-    }
-
-    if (expectedProbabilities.size() > 1) {
-      assertThat(new GTest().gTest(expectedValues, observedValues)).isGreaterThan(0.01);
-    } else {
-      assertThat((double) observedValues[0]).isEqualTo(expectedValues[0]);
-    }
+    return sb.toString();
   }
 }
